+++ src/render/index.ts
@@ -8,6 +8,7 @@ export * from './fighter-animation-view';
 export * from './background-scene';
 export * from './fight-presentation';
 export * from './arcade-runtime-contract';
+export * from './arcade-runtime-adapter';
 
 export * from './vfx';
 export * from './sprites';